+++ qiskit/tools/jupyter/jupyter_magics.py
@@ -12,9 +12,14 @@ import threading
 from IPython.display import display                              # pylint: disable=import-error
 from IPython.core import magic_arguments                         # pylint: disable=import-error
 from IPython.core.magic import cell_magic, Magics, magics_class  # pylint: disable=import-error
-import ipywidgets as widgets                                     # pylint: disable=import-error
+try:
+    import ipywidgets as widgets           # pylint: disable=import-error
+except ImportError:
+    raise ImportError('These functions  need ipywidgets. '
+                      'Run "pip install ipywidgets" before.')
 import qiskit
-from .progressbar import HTMLProgressBar, TextProgressBar
+from qiskit.tools.events.progressbar import TextProgressBar
+from .progressbar import HTMLProgressBar
 
 
 def _html_checker(job_var, interval, status, header):