+++ qiskit/tools/visualization/__init__.py
@@ -22,7 +22,6 @@ from qiskit.tools.visualization._state_visualization import (plot_state_hinton,
 from ._circuit_visualization import circuit_drawer, plot_circuit, generate_latex_source, \
     latex_circuit_drawer, matplotlib_circuit_drawer, _text_circuit_drawer, qx_color_scheme
 from ._error import VisualizationError
-from ._dag_visualization import dag_drawer
 from ._matplotlib import HAS_MATPLOTLIB
 
 if ('ipykernel' in sys.modules) and ('spyder' not in sys.modules):
@@ -34,3 +33,6 @@ if ('ipykernel' in sys.modules) and ('spyder' not in sys.modules):
                                                             iplot_state_hinton,
                                                             iplot_histogram,
                                                             iplot_state_paulivec)
+
+if HAS_MATPLOTLIB:
+    from ._gate_map import plot_gate_map