+++ qiskit/transpiler/passes/__init__.py
@@ -9,7 +9,10 @@
 
 from .cx_cancellation import CXCancellation
 from .fixed_point import FixedPoint
+from .optimize_1q_gates import Optimize1qGates
+from .decompose import Decompose
 from .mapping.check_map import CheckMap
 from .mapping.basic_mapper import BasicMapper
 from .mapping.direction_mapper import DirectionMapper
 from .mapping.unroller import Unroller
+from .mapping.lookahead_mapper import LookaheadMapper