+++ CHANGELOG.rst
@@ -43,6 +43,8 @@ Added
   QuantumCircuit class. (#1172)
 - New methods in QuantumCircuit for common circuit metrics:
   `size()`, `depth()`, `width()`, `count_ops()`, `num_tensor_factors()` (#1285)
+- Added `backend_monitor` and `backend_overview` Jupyter magics,
+  as well as `plot_coupling_map` (#1231)
 - Added a `Layout` object (#1313)
 - Added a BasicMapper pass (#1270)
 - New `plot_bloch_multivector()` to plot Bloch vectors from a tensored state
@@ -54,6 +56,7 @@ Added
 - Added a ``qiskit.converters`` module for translation between commonly used
   representations of a circuit: ``dag_to_circuits``, ``circuits_to_dag``,
   ``qobj_to_circuits``, ``circuits_to_qobj``, ``ast_to_dag``.
+- Added lookahead mapper as new transpiler pass (#1140).
 
 Changed
 """""""
@@ -97,6 +100,11 @@ Changed
   processing is only done internally (#1397).
 - Moved all the circuit modules into a circuit module but for most users it is still 
   imported in the top level for QuantumCircuit, QuantumRegister, ClassicalRegister
+- Breaking change: ``qiskit.backends.aer`` has been removed in favour of
+  ``qiskit.backends.builtinsimulators`` (Python simulators) and
+  ``qiskit.backends.legacysimulators`` (C++ simulators) (#1484)
+- Breaking change: ``Aer`` in ``qiskit`` root module has been removed. Instead ``Simulators`` will
+  stand for Python simulators and ``LegacySimulators`` for those in C++. (#1484)
 
 
 Deprecated